requirements.txt
numpy==1.11.2
scipy==0.18.1
BTrees==4.3.1
DoubleMetaphone==0.1
PyLBFGS==0.2.0.3
PyYAML==3.12
Pygments==2.1.3
SQLAlchemy==1.0.13
affinegap==1.10
argparse==1.2.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
boto3==1.4.1
botocore==1.4.92
canonicalize==1.3
categorical-distance==1.9
certifi==2016.9.26
collate==0.1.0
cycler==0.10.0
decorator==4.0.10
dedupe==1.4.15
dedupe-hcluster==0.3.2
docutils==0.13.1
enum34==1.1.6
fastcluster==1.1.22
future==0.16.0
futures==3.0.5
haversine==0.4.5
highered==0.2.1
ipykernel==4.3.1
ipython==5.1.0
ipython-genutils==0.1.0
jmespath==0.9.0
jupyter-client==4.4.0
jupyter-core==4.2.1
matplotlib==1.5.3
numexpr==2.6.1
pandas==0.19.1
pathlib2==2.1.0
persistent==4.2.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.1
ptyprocess==0.5.1
pyhacrf-datamade==0.2.0
pymaybe==0.1.6
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
requests==2.10.0
rlr==2.4
s3transfer==0.1.10
scikit-learn==0.17.1
seaborn==0.7.1
simplecosine==1.1
simplegeneric==0.8.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
tables==3.3.0
tornado==4.4.2
tqdm==4.8.3
traitlets==4.3.1
wcwidth==0.1.7
wsgiref==0.1.2
zope.index==4.2.0
zope.interface==4.3.3